requirements.txt
# === FICHIER : requirements.txt (VERSION CORRIGÉE) ===
# Ordre important : numpy en premier pour éviter les conflits de wheels
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2
xgboost==2.1.1
mlflow==2.16.2
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
matplotlib==3.9.2
seaborn==0.13.2
imbalanced-learn==0.12.4
pytest==8.3.3
black==24.8.0
flake8==7.1.1
pandera==0.20.1
pyyaml==6.0.2
requests==2.32.3
joblib==1.4.2
scipy==1.14.1
great-expectations==0.18.3